chore(cmd): Remove IntVar and move to Int64Var

+++ internal/cmd/common/flags.go
@@ -72,7 +72,7 @@ func PopulateCommonFlags(c *base.Command, f *base.FlagSet, resourceType string, 
 				Usage:  fmt.Sprintf("Description to set on the %s.", resourceType),
 			})
 		case "version":
-			f.IntVar(&base.IntVar{
+			f.Int64Var(&base.Int64Var{
 				Name:   "version",
 				Target: &c.FlagVersion,
 				Usage:  fmt.Sprintf("The version of the %s against which to perform an update operation. If not specified, the command will perform a check-and-set automatically.", resourceType),
